Extract interfaces related to language pairs to their own file
